FIX: required big int field as id

+++ src/entity/message.py
@@ -5,7 +5,7 @@ from entity.user import User
 
 
 class Message(BaseEntity):
-    msg_id = peewee.IntegerField(primary_key=True)
+    msg_id = peewee.BigIntegerField(primary_key=True)
     text = peewee.TextField()
     chat_id = peewee.BigIntegerField()
     user = peewee.ForeignKeyField(model=User)
